monitor/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
brotli==1.2.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
cssselect2==0.7.0
Django==4.2.26
djangorestframework==3.15.2
fonttools==4.57.0
html5lib==1.1
idna==3.11
pillow==10.4.0
pkg_resources==0.0.0
psutil==7.1.3
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
pyphen==0.16.0
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
typing_extensions==4.13.2
urllib3==2.2.3
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3.post1
